tg_bot/bot_requirements.txt
aiocron==1.8
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
croniter==2.0.5
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
